Move `stake_pool_store` tests in `tests` module
