+++ mithril-signer/src/database/tests/mod.rs
@@ -1,1 +1,2 @@
 mod protocol_initializer;
+mod stake_pool;
